+++ CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Upcoming changes...
 
+## [1.5.1] - 2023-04-21
+### Added
+- Added support scanning/fingeprinting file contents from STDIN
+  - `cat test.py | scanoss-py scan --stdin test.py -o results.json`
+  - `cat test.py | scanoss-py wfp --stdin test.py -o fingers.wfp`
+
 ## [1.5.0] - 2023-03-21
 ### Added
 - Added support for component cryptographic reporting
@@ -224,3 +230,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [1.4.0]: https://github.com/scanoss/scanoss.py/compare/v1.3.7...v1.4.0
 [1.4.2]: https://github.com/scanoss/scanoss.py/compare/v1.4.0...v1.4.2
 [1.5.0]: https://github.com/scanoss/scanoss.py/compare/v1.4.2...v1.5.0
+[1.5.1]: https://github.com/scanoss/scanoss.py/compare/v1.5.0...v1.5.1